+++ bindings/nodejs/lib/bindings.ts
@@ -5,13 +5,7 @@ import type { MessageHandler } from './MessageHandler';
 // @ts-ignore: path is set to match runtime transpiled js path
 import addon = require('../../build/Release/index.node');
 
-const {
-    initLogger,
-    sendMessage,
-    messageHandlerNew,
-    listen,
-    destroy,
-} = addon;
+const { initLogger, sendMessage, messageHandlerNew, listen, destroy } = addon;
 
 const sendMessageAsync = (
     message: string,