+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Default` to `FilterOptions`;
 - `ParticipationEventRegistrationOptions` type;
+- `AccountManager::update_node_auth()`;
+- `Message::UpdateNodeAuth`;
 
 ### Changed